HW7: Added behaviours

+++ Dotnet.Homeworks.Features/UserManagement/Commands/DeleteUserByAdmin/DeleteUserByAdminCommandHandler.cs
@@ -1,19 +1,25 @@
-﻿using Dotnet.Homeworks.Infrastructure.Cqrs.Commands;
+﻿using Dotnet.Homeworks.Domain.Abstractions.Repositories;
+using Dotnet.Homeworks.Infrastructure.Cqrs.Commands;
 using Dotnet.Homeworks.Infrastructure.UnitOfWork;
 using Dotnet.Homeworks.Shared.Dto;
+using Microsoft.EntityFrameworkCore;
 
 namespace Dotnet.Homeworks.Features.UserManagement.Commands.DeleteUserByAdmin;
 
 public class DeleteUserByAdminCommandHandler : ICommandHandler<DeleteUserByAdminCommand>
 {
-    public DeleteUserByAdminCommandHandler(UnitOfWork unitOfWork)
+    public DeleteUserByAdminCommandHandler(IUnitOfWork unitOfWork, IUserRepository userRepository)
     {
         UnitOfWork = unitOfWork;
+        UserRepository = userRepository;
     }
 
-    private UnitOfWork UnitOfWork { get; }
+    private IUnitOfWork UnitOfWork { get; }
+    private IUserRepository UserRepository { get; }
     public async Task<Result> Handle(DeleteUserByAdminCommand request, CancellationToken cancellationToken)
     {
+        if (!(await (await UserRepository.GetUsersAsync(cancellationToken)).AnyAsync(user => user.Id == request.Guid, cancellationToken: cancellationToken)))
+            return new Result(false, "No such user found.");
         await UnitOfWork.UserRepository.DeleteUserByGuidAsync(request.Guid, cancellationToken);
         await UnitOfWork.SaveChangesAsync(cancellationToken);
         return new Result(true);
